Update CLI interface, Use Tika for reading, Platform independent coding
# Update CLI interface
This is done to add new features, remove unhelpful features, provided more consistency and ease of use to the end user 

# Use Tika for reading
After lots of testing and reading, Tika seemed to be the best option and all in one solution to convert any document to text. Hence new functions were added which used Tika for reading files/documents

# Platform independent coding
The code is being written such that it can quickly be ported to run on Windows and Mac as well.